+++ Runtime/Scripts/Utilities/Tweening/CURVE.cs
@@ -1,0 +1,11 @@
+﻿namespace Utilities.Tweening
+{
+    public enum CURVE
+    {
+        LINEAR,
+        EASE_IN,
+        EASE_OUT,
+        EASE_IN_OUT,
+        
+    }
+}