+++ Samples~/MainMenu/MainMenuUI.cs
@@ -1,4 +1,5 @@
 using UnityEngine;
+using UnityEngine.Assertions;
 using UnityEngine.SceneManagement;
 using UnityEngine.UI;
 using Utilities;
@@ -7,6 +8,7 @@ namespace UI
 {
     public class MainMenuUI : MonoBehaviour
     {
+        [Header("Main Menu")]
         [SerializeField]
         private Button playButton;
         [SerializeField]
@@ -14,13 +16,19 @@ namespace UI
         [SerializeField]
         private Button quitButton;
 
+        [SerializeField, Header("Windows")] 
+        private BaseUIWindow settingsWindow;
         //============================================================================================================//
         
         // Start is called before the first frame update
         private void Start()
         {
+            Assert.IsNotNull(settingsWindow);
+            
             ScreenFader.ForceSetColorBlack();
             playButton.onClick.AddListener(OnPlayButtonPressed);
+            
+            settingsButton.onClick.AddListener(OnSettingButtonPressed);
 
 #if UNITY_WEBGL
             quitButton.gameObject.SetActive(false);
@@ -39,6 +47,11 @@ namespace UI
             {
                 SceneManager.LoadScene(1);
             });
+        }
+        
+        private void OnSettingButtonPressed()
+        {
+            settingsWindow.OpenWindow();
         }
 
         private void OnQuitButtonPressed()